refactor(api): namespace v1 and v2 packages
